+++ src/GitHub_Updater/Traits/Basic_Auth_Loader.php
@@ -15,6 +15,8 @@ use Fragen\GitHub_Updater\Install;
 use Fragen\GitHub_Updater\API\Bitbucket_API;
 use Fragen\GitHub_Updater\API\Bitbucket_Server_API;
 use Fragen\GitHub_Updater\API\GitHub_API;
+use Fragen\GitHub_Updater\API\GitLab_API;
+use Fragen\GitHub_Updater\API\Gitea_API;
 
 /*
  * Exit if called directly.
@@ -32,7 +34,7 @@ trait Basic_Auth_Loader {
 	 *
 	 * @var array
 	 */
-	private static $basic_auth_required = [ 'Bitbucket', 'GitHub' ];
+	private static $basic_auth_required = [ 'Bitbucket', 'GitHub', 'GitLab', 'Gitea' ];
 
 	/**
 	 * Load hooks for Bitbucket authentication headers.
@@ -41,7 +43,6 @@ trait Basic_Auth_Loader {
 	 */
 	public function load_authentication_hooks() {
 		add_filter( 'http_request_args', [ $this, 'maybe_basic_authenticate_http' ], 5, 2 );
-		add_filter( 'http_request_args', [ $this, 'maybe_token_authenticate_http' ], 5, 2 );
 		add_filter( 'http_request_args', [ $this, 'http_release_asset_auth' ], 15, 2 );
 	}
 
@@ -52,7 +53,6 @@ trait Basic_Auth_Loader {
 	 */
 	public function remove_authentication_hooks() {
 		remove_filter( 'http_request_args', [ $this, 'maybe_basic_authenticate_http' ] );
-		remove_filter( 'http_request_args', [ $this, 'maybe_token_authenticate_http' ] );
 		remove_filter( 'http_request_args', [ $this, 'http_release_asset_auth' ] );
 	}
 
@@ -69,32 +69,23 @@ trait Basic_Auth_Loader {
 	 */
 	public function maybe_basic_authenticate_http( $args, $url ) {
 		$credentials = $this->get_credentials( $url );
-		if ( 'bitbucket' === $credentials['type'] && $credentials['private'] && $credentials['isset'] && ! $credentials['api.wordpress'] ) {
+		if ( ! $credentials['isset'] || $credentials['api.wordpress'] ) {
+			return $args;
+		}
+		if ( 'bitbucket' === $credentials['type'] ) {
 			$username = $credentials['username'];
 			$password = $credentials['password'];
 
 			// phpcs:ignore WordPress.PHP.DiscouragedPHPFunctions.obfuscation_base64_encode
 			$args['headers']['Authorization'] = 'Basic ' . base64_encode( "$username:$password" );
 		}
-		remove_filter( 'http_request_args', [ $this, 'maybe_basic_authenticate_http' ] );
-
-		return $args;
-	}
-
-	/**
-	 * Add Basic Authentication $args to http_request_args filter hook.
-	 * Adds `token` header for GitHub API.
-	 *
-	 * @param array  $args Args passed to the URL.
-	 * @param string $url  The URL.
-	 *
-	 * @return array $args
-	 */
-	public function maybe_token_authenticate_http( $args, $url ) {
-		$credentials = $this->get_credentials( $url );
-		if ( 'github' === $credentials['type'] && $credentials['isset'] && ! $credentials['api.wordpress'] ) {
+		if ( 'github' === $credentials['type'] || 'gitea' === $credentials['type'] ) {
 			$args['headers']['Authorization'] = 'token ' . $credentials['token'];
 		}
+		if ( 'gitlab' === $credentials['type'] ) {
+			$args['headers']['Authorization'] = 'Bearer ' . $credentials['token'];
+		}
+
 		remove_filter( 'http_request_args', [ $this, 'maybe_basic_authenticate_http' ] );
 
 		return $args;
@@ -122,7 +113,7 @@ trait Basic_Auth_Loader {
 			'token'         => null,
 			'type'          => null,
 		];
-		$hosts        = [ 'bitbucket.org', 'api.bitbucket.org', 'github.com', 'api.github.com' ];
+		$hosts        = [ 'bitbucket.org', 'api.bitbucket.org' ];
 
 		$repos = array_merge(
 			Singleton::get_instance( 'Plugin', $this )->get_plugin_configs(),
@@ -142,15 +133,27 @@ trait Basic_Auth_Loader {
 				break;
 			case 'github':
 			case $type instanceof GitHub_API:
-				$github_com = in_array( $headers['host'], $hosts, true );
-				$token      = self::$options['github_access_token'];
-				$type       = 'github';
+				$token = ! empty( self::$options[ $slug ] ) ? self::$options[ $slug ] : self::$options['github_access_token'];
+				$type  = 'github';
 				break;
+			case 'gitlab':
+			case $type instanceof GitLab_API:
+				$token = ! empty( self::$options[ $slug ] ) ? self::$options[ $slug ] : self::$options['gitlab_access_token'];
+				$type  = 'gitlab';
+				break;
+			case 'gitea':
+			case $type instanceof Gitea_API:
+				$token = ! empty( self::$options[ $slug ] ) ? self::$options[ $slug ] : self::$options['gitea_access_token'];
+				$type  = 'gitea';
 		}
 
 		// TODO: can use `( $this->caller )::$options` in PHP7.
 		$caller          = $this->get_class_vars( 'Base', 'caller' );
 		static::$options = $caller instanceof Install ? $caller::$options : static::$options;
+
+		if ( ! $slug || $credentials['api.wordpress'] || $type !== $repos[ $slug ]->git ) {
+			return $credentials;
+		}
 
 		if ( 'bitbucket' === $type && isset( static::$options[ $username_key ], static::$options[ $password_key ] ) ) {
 			$credentials['username'] = static::$options[ $username_key ];
@@ -160,7 +163,7 @@ trait Basic_Auth_Loader {
 			$credentials['type']     = $type;
 		}
 
-		if ( 'github' === $type ) {
+		if ( 'github' === $type || 'gitlab' === $type || 'gitea' === $type ) {
 			$credentials['isset']   = true;
 			$credentials['private'] = $this->is_repo_private( $url );
 			$credentials['type']    = $type;
@@ -182,6 +185,7 @@ trait Basic_Auth_Loader {
 	private function get_slug_for_credentials( $headers, $repos, $url ) {
 		$slug = isset( $_REQUEST['slug'] ) ? $_REQUEST['slug'] : false;
 		$slug = ! $slug && isset( $_REQUEST['plugin'] ) ? $_REQUEST['plugin'] : $slug;
+		$slug = false !== strpos( $slug, '/' ) ? dirname( $slug ) : $slug;
 		$slug = ! $slug && isset( $_REQUEST['theme'] ) ? $_REQUEST['theme'] : $slug;
 
 		// Set for bulk upgrade.